+++ ForumTask.BLL/Interfaces/IMarkService.cs
@@ -19,7 +19,7 @@ namespace ForumTask.BLL.Interfaces
         /// <param name="userId">User which mark to get</param>
         /// <param name="messageId">Message from which mark to get</param>
         /// <returns>Value of mark or 0 if don`t exist</returns>
-        Task GetOwnAsync(int userId, long messageId);
+        Task GetOwnAsync(long userId, long messageId);
 
         /// <summary>
         /// Gets total mark of message with <paramref name="messageId"/>
